+++ language_server/test/ws_mngr_mock.h
@@ -99,6 +99,11 @@ public:
         branch_information,
         (const char* document_uri, workspace_manager_response<continuous_sequence<branch_info>> resp),
         (override));
+
+    MOCK_METHOD(void,
+        folding,
+        (const char* document_uri, workspace_manager_response<continuous_sequence<folding_range>> resp),
+        (override));
 };
 
 } // namespace hlasm_plugin::language_server::test