+++ parser_library/src/workspaces/workspace.h
@@ -32,6 +32,7 @@
 #include "debugging/debugger_configuration.h"
 #include "diagnosable_impl.h"
 #include "file_manager_vfm.h"
+#include "folding_range.h"
 #include "lib_config.h"
 #include "macro_cache.h"
 #include "message_consumer.h"
@@ -115,6 +116,8 @@ public:
     std::vector<token_info> semantic_tokens(const resource_location& document_loc) const;
 
     std::vector<branch_info> branch_information(const resource_location& document_loc) const;
+
+    std::vector<folding_range> folding(const resource_location& document_loc) const;
 
     std::optional<performance_metrics> last_metrics(const resource_location& document_loc) const;